types: add support for enable_ipv6
The enable_ipv6 property specified in the network spec was missing
from the NetworkConfig struct.  Because of this ommission, users of
this library had no visibility into this setting.  In particular,
this prevented the docker-compose v2 application from creating a
Network object in docker with an appropriate setting for the
EnableIPv6 option.

This change parses the "enable_ipv6" property into the NetworkConfig
struct and makes it available in the EnableIPv6 field for any
downstream library users to use.

Signed-off-by: Stacey Sheldon <stac@es.net>