Added support for volume bind option SELinux :z :Z
The docker-compose CLI v2 uses the docker/compose-go package for
parsing Docker Compose YAML specification files and currently it misses
support for the SELinux relabeling with `:z` or `:Z`.

It is a regression compared to docker-compose CLI v1.

This feature allow to use the SELinux field to set the `:z` or `:Z`
bind option for relabeling SELinux label.

References:
- https://docs.docker.com/storage/bind-mounts/#configure-the-selinux-label

Signed-off-by: Tymoteusz Blazejczyk <tymoteusz.blazejczyk@tymonx.com>