Fix inherited env var not found in multi-var .env
The package `dotenv` implements inherited environment variables which
allows variables to be set on the host and passed to a container using
key-only pairs in a .env file. Package `compose-go` uses this feature by
calling the function `ParseWithLookup` of `dotenv`.

One expectation of this feature is that an environment variable will not
be passed to the container if the host does not set it. This scenario is
tested in `dotenv`. However, there is an issue if a `.env` file contains
multiple key pairs resulting of environment variables not being set
properly in the container.

This commit fixes this issue.

Signed-off-by: Frederic Lemay <461102+flemay@users.noreply.github.com>