env: fix errors on valid interpolation expressions (#307)
The parser was too strict here and would reject valid constructs
that used an unescaped `$` that was not part of a variable
expression (and not ambiguous).

Now, only errors for unmatched braced expressions (e.g. `${FOO`)
are returned, but other valid cases are ignored and the `$` will
be treated literally, e.g. `a $ string` -> `a $ string`, which is
the same as in POSIX.

Signed-off-by: Milas Bowman <milas.bowman@docker.com>