Fix options_windows_test using valid working dir
After applying package name constraints uniformly in the previous
commit, cli/options_windows_test broke on Windows with:

```
--- FAIL: TestConvertWithEnvVar (0.00s)                                                                                                                              options_windows_test.go:35: assertion failed: error is not nil: "\\" is not a valid project name...
```

This is because the `TestConvertWithEnvVar` test was setting the working
directory as `C://`. No other mechanism defined the package name,
causing the non-normalized package name to be `/` and the normalized
package name to be ``.

Technically, this is exactly what we want: The root directory
_shouldn't_ be used for the project name, which will be empty. This
violates the new `^[a-z0-9][a-z0-9_-]*$` package name constraint.

As a result, the solution is to update the working directory to
`C:/working-dir/` and to adjust the expected results accordingly.

Signed-off-by: Mike Bland <mbland@acm.org>