Test invalid project names: empty, root directory
Since the spec now explicitly rejects empty project names, it's worth
explicitly testing for this case. Since the root directory will
normalize as the empty string, it's worth specifying this requirement
explicitly in a test case as well. (This is why the options_window_test
started failing two commits ago, though that wasn't the intent of that
test.)

added an extra check to loader.CheckOriginalProjectNameIsNormalized
specifically for empty project names.

Signed-off-by: Mike Bland <mbland@acm.org>