cli: allow setting empty name in options (to infer)
The `ProjectOptions` includes config to pass to the loader,
but most of the fields, including `Name`, are optional. If
unset/empty, the loader will determine a project name (if
possible).

Allow (re-)setting it to empty here. This is mostly for
convenience based on the way these helpers are used in
practice, so that it's possible to pass a value from e.g.
a CLI flag in, which will be empty if unset.

Also added a bunch of documentation to the type.

Signed-off-by: Milas Bowman <milas.bowman@docker.com>