loader: fix overly aggressive project name validation
There were cases where the loader would infer a perfectly valid
project name by normalizing a directory, but would then re-validate
the un-normalized project directory and throw an error.

This moves all the validation into a single spot during the load,
which is the only point that it has all the context.

Some of this logic needs to be pushed out of here: it should NOT
be reading the config files, for example.

For now, this resolves the main issues.

Signed-off-by: Milas Bowman <milas.bowman@docker.com>