loader: report project structure
With `include`'d Compose files, it's important to know the origin
of each service to be able to properly resolve paths to them,
for example.

A new `FileMeta` field on the `Project` type provides detailed
access to the underlying file structure/hierarchy.

This replaces `IncludeReferences` and `ComposeFiles`, which were
inconsistent and not comprehensive.

Signed-off-by: Milas Bowman <milas.bowman@docker.com>