networks.enable_ipv6: make it optional
This field isn't marked as required in the compose-spec, but it's
implemented as a basic `bool` in this package. This means it
automatically defaults to `false` when not specified.

A recent change was made to Docker Engine to allow users to define a
daemon-wide default value for the equivalent API field. Without this
change, Compose would always send a value to the Engine API, preventing
it from defining its own default value.

See https://github.com/moby/moby/pull/47867.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>